version asserts and mitigation hash function unit tests

+++ src/core.h
@@ -160,4 +160,4 @@ void b3JoinThread( b3Thread* t );
 
 void b3StrCpy( char* dst, int size, const char* src );
 
-uint64_t b3Hash64NonZero( const uint8_t* bytes, size_t n );
+uint64_t b3Hash64NonZero( const uint8_t* bytes, int n );
